+++ test/auth.spec.js
@@ -38,7 +38,7 @@ function stubTokenRequest() {
         });
 }
 
-function stubTokenRequestError() {
+function stubTokenRequestError()    {
     nock('https://oauth2.googleapis.com')
         .post('/token')
         .reply(400, {
@@ -139,7 +139,7 @@ describe('UserAuthorizer', function() {
                 const credentials = authorizer.getUserCredentials(
                     'user@example.com',
                     'https://www.googleapis.com/auth/slides');
-                return expect(credentials).to.eventually.have.deep.property('credentials.access_token', 'new_token');
+                return expect(credentials).to.eventually.have.nested.property('credentials.access_token', 'new_token');
             });
         });
 
@@ -147,14 +147,14 @@ describe('UserAuthorizer', function() {
             it('should return token if still current', function() {
                 const authorizer = new UserAuthorizer(options);
                 const credentials = authorizer.getUserCredentials('current', 'https://www.googleapis.com/auth/slides');
-                return expect(credentials).to.eventually.have.deep.property('credentials.access_token', 'ya29.123');
+                return expect(credentials).to.eventually.have.nested.property('credentials.access_token', 'ya29.123');
             });
 
             it('should refresh token if expired', function() {
                 stubTokenRequest();
                 const authorizer = new UserAuthorizer(options);
                 const credentials = authorizer.getUserCredentials('expired', 'https://www.googleapis.com/auth/slides');
-                return expect(credentials).to.eventually.have.deep.property('credentials.access_token', 'new_token');
+                return expect(credentials).to.eventually.have.nested.property('credentials.access_token', 'new_token');
             });
         });
     });